+++ py-polars/tests/unit/io/test_scan.py
@@ -9,7 +9,6 @@ from typing import TYPE_CHECKING, Any, Callable
 import pytest
 
 import polars as pl
-from polars.exceptions import ComputeError
 from polars.testing.asserts.frame import assert_frame_equal
 
 if TYPE_CHECKING:
@@ -586,25 +585,4 @@ def test_scan_nonexistent_path(format: str) -> None:
 
     # Upon collection, it should fail
     with pytest.raises(FileNotFoundError):
-        result.collect()
-
-
-@pytest.mark.slow()
-@pytest.mark.parametrize("format", ["parquet", "csv", "ndjson", "ipc"])
-def test_scan_nonexistent_cloud_path_17444(format: str) -> None:
-    # https://github.com/pola-rs/polars/issues/17444
-
-    path_str = f"s3://my-nonexistent-bucket/data.{format}"
-    scan_function = getattr(pl, f"scan_{format}")
-
-    # Just calling the scan function should not raise any errors
-    if format == "ndjson":
-        # NDJSON does not have a `retries` parameter yet - so use the default
-        result = scan_function(path_str)
-    else:
-        result = scan_function(path_str, retries=0)
-    assert isinstance(result, pl.LazyFrame)
-
-    # Upon collection, it should fail
-    with pytest.raises(ComputeError):
         result.collect()
